refactor: image bucket

+++ pages/api/posts/post.ts
@@ -1,12 +1,14 @@
 import { withApiAuth } from '@supabase/auth-helpers-nextjs';
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { v4 } from 'uuid';
 
 import { prisma } from '@/utils/db';
+import { imageKit } from '@/utils/imageKit';
 
 const handler = withApiAuth(async (req: NextApiRequest, res: NextApiResponse) => {
   const {
     method,
-    body: { description, location, postID, type, publicUrl, author_id, uuid, post_id },
+    body: { description, location, postID, type, author_id, post_id, imageFile },
   } = req;
 
   if (method !== 'POST') {
@@ -32,12 +34,18 @@ const handler = withApiAuth(async (req: NextApiRequest, res: NextApiResponse) =>
 
   if (type === 'CREATE') {
     try {
+      const uuid = v4();
+      const uploadedImage = await imageKit.upload({
+        file: imageFile,
+        fileName: `post.webp`,
+        folder: `${author_id}/posts/${uuid}`,
+      });
       await prisma.posts.create({
         data: {
-          img: publicUrl,
-          description: description,
-          author_id: author_id,
-          img_uuid: uuid,
+          img: uploadedImage.url,
+          file_id: uploadedImage.fileId,
+          description,
+          author_id,
           location,
         },
         select: {
@@ -45,14 +53,26 @@ const handler = withApiAuth(async (req: NextApiRequest, res: NextApiResponse) =>
           id: true,
         },
       });
-      res.status(200).send('success');
+      res.status(200).send(200);
     } catch (e) {
-      res.status(400);
+      res.status(400).send(400);
     }
   }
 
   if (type === 'REMOVE') {
     try {
+      const deletedPost = await prisma.posts.delete({
+        where: {
+          id: post_id,
+        },
+      });
+
+      if (!deletedPost.file_id) {
+        res.status(400).send(400);
+        return;
+      }
+
+      await imageKit.deleteFile(deletedPost.file_id);
       await prisma.posts_likes.deleteMany({
         where: {
           post_id,
@@ -63,17 +83,20 @@ const handler = withApiAuth(async (req: NextApiRequest, res: NextApiResponse) =>
           post_id,
         },
       });
-      await prisma.posts.delete({
-        where: {
-          id: post_id,
-        },
-      });
 
-      res.status(200).send('success');
+      res.status(200).send(200);
     } catch (e) {
-      res.status(400);
+      res.status(400).send(400);
     }
   }
 });
 
 export default handler;
+
+export const config = {
+  api: {
+    bodyParser: {
+      sizeLimit: '12mb',
+    },
+  },
+};
